fixes1255: crash fix - purge every remaining %u/%x/%lu/%lx from macsurf_debug_log_writef
Root cause of the home-page crash-loop reported after fixes1254: macsurf_debug_log_writef's
hand-rolled formatter (fmt_vformat, macsurf_debug_log.c) only recognizes %d, %ld, %p, %s,
%% - an unrecognized specifier prints its literal characters WITHOUT calling va_arg, so it
never consumes that argument. Every subsequent %s/%d/%p in the same call then reads the
wrong slot, shifted by one. fixes1254's four new SPIPE lines all used %u for the ordinal,
immediately followed by %s (type/url) - the ordinal's small integer value (1, 2, 3...) got
reinterpreted as a char* and dereferenced by fmt_append_str, crashing on literally every
page load with a script tag (i.e. every page, including the plain home page). Fixed all
four (plus fixes1253's SCRIPTFETCHFAIL, same bug) by switching to %ld with (long) casts,
matching this file's own established convention (see fixes717/907/927/160b comments already
in the tree for the exact same bug class).

Also swept the rest of the codebase for the same landmine, scoped to files that actually
compile into MacSurf (excluded riscos/amiga/windows/atari frontends - never built here).
Found and fixed 7 more live instances, none of which were crashing (all trailing/no %s
downstream so they only silently misprinted, not corrupted alignment) but same latent bug
class: macos9_svg_inline.c (%08x x2), plotters.c (%08lx x2, LIFE-prefixed blend diagnostics),
script.c's deferred_unpause (%x), macos9_image.c (%lu), box_construct.c's svg_box trace
(%x), and three WORK-gated timer diagnostics in macsurf_qjs.c (%lu x2 each). All converted
to %ld + (long) casts, this formatter has no hex support so none can preserve hex display.

+++ browser/netsurf/content/handlers/html/html_css.c
@@ -339,8 +339,8 @@ static void html_css_process_modified_styles(void *pw)
 		} else {
 			macsurf_debug_log_writef(
 				"LIFE css modified stylesheet retry cap hit content=%p "
-				"count=%u",
-				(void *)c, retry_count);
+				"count=%ld",
+				(void *)c, (long) retry_count);
 			for (i = 0, s = c->stylesheets;
 			     i != c->stylesheet_count; i++, s++) {
 				s->modified = false;
